Acquirer Queue
Add optional AcquirerId to AcquireOptions.  If passed, it will hold the
acquirer's place in line if the requested lock is already held.

When the lock expires or is released, the next acquirer will be the one
that has been waiting the longest.

When a lease is released, if there are Queue Members, expire it so it
can be acquired right away.

Signed-off-by: Christian G. Warden <cwarden@xerus.org>